requirements.txt
djangorestframework
django>=2.1
Pillow==6.1.0